Add Types & Rename & Base for Physic & Many other things
Add PUBLIC CompileDefinition VXM_FLOAT/VXM_DOUBLE
Add the Real Type and implement Vec3, etc depending on a macro VXM_FLOAT/VXM_DOUBLE
Add an overlay for the Math library by using glm bellow my own implementation of Math.
Use the new Vec3 type in the ImGuiLib
Rename the "ToString" of the event into "string" (which is a breaking change...)
Add Particle for the physic. Using simple Euler Integration for now.
Add PrimitiveComponent to be able to add simple quad & cube in the scene and tests things out (link the lightning later on and the physic currently)
Probably others things that I've forgotten...