[WIP] - Converting generic Shader to two distinct class
Starting the conversion of the Shader class to Graphics Shader & Compute Shader.

Thoses will be derived to their OpenGL version that will inherit from a specific OpenGLShader and their respective base class.

Objectifs is to be able to use them in a "Pipeline".
So need to define the Input (and Output for ComputeShader) so that it can be used.

Later on will be use inside RenderAction for a RenderPipeline that is in order set in the necessary objects (like a mesh, a model, or a Terrain)