Create basics for a ComputeShaderAction + upgrade a bit the pipeline
Welp, just create a default action for my compute shader, fill the pipeline do pretend to do things.
Still need to add all the action, make an asset, use it in some componenet, try to make a minecraft-like terrain using the points and the compute shader and shaders and all.